Add SourcesLibWebRTC.txt https://bugs.webkit.org/show_bug.cgi?id=247827
Reviewed by Michael Catanzaro.

Move libwebrtc files into a separate unified sources listing. Modify
CMake and XCode build accordingly. Move LibWebRTCVPX* files into
`platform/libwebrtc`. Fix any build issues that cropped up with the
unified build, such as renaming the `vpxQueue` functions so the symbol
isn't multiply defined.

* Source/WebCore/CMakeLists.txt:
* Source/WebCore/Scripts/generate-unified-sources.sh:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/platform/GStreamer.cmake:
* Source/WebCore/platform/SourcesLibWebRTC.txt: Added.
* Source/WebCore/platform/audio/glib/MediaSessionManagerGLib.cpp:
* Source/WebCore/platform/libwebrtc/LibWebRTCVPXVideoDecoder.cpp: Renamed from Source/WebCore/platform/LibWebRTCVPXVideoDecoder.cpp.
* Source/WebCore/platform/libwebrtc/LibWebRTCVPXVideoDecoder.h: Renamed from Source/WebCore/platform/LibWebRTCVPXVideoDecoder.h.
* Source/WebCore/platform/libwebrtc/LibWebRTCVPXVideoEncoder.cpp: Renamed from Source/WebCore/platform/LibWebRTCVPXVideoEncoder.cpp.
* Source/WebCore/platform/libwebrtc/LibWebRTCVPXVideoEncoder.h: Renamed from Source/WebCore/platform/LibWebRTCVPXVideoEncoder.h.

Canonical link: https://commits.webkit.org/256690@main
