+++ Source/WebCore/platform/audio/glib/MediaSessionManagerGLib.cpp
@@ -29,8 +29,8 @@
 #include "NowPlayingInfo.h"
 #include "PlatformMediaSession.h"
 #include "PlatformStrategies.h"
-
 #include <gio/gio.h>
+#include <wtf/glib/GUniquePtr.h>
 
 // https://specifications.freedesktop.org/mpris-spec/latest/
 static const char s_mprisInterface[] =
